requirements.txt
httpx==0.27.0
loguru==0.7.0
python-dotenv==1.0.0
pydantic==2.8.2
aiogram